Fix broken info-diffs for groups of two packages
When %PackageInfo array has exactly two keys, it is assumed the package
has been renamed. That is not always the case, for example when there is
a group of two packages (same in both old and new version).
This fix prevents that by adding another contition to test if neither
package name is present in both versions simultaneously.